+++ sc-memory/sc-core/sc_helper.h
@@ -76,10 +76,4 @@ sc_helper_resolve_system_identifier(sc_memory_context const * ctx, const char * 
  */
 _SC_EXTERN sc_bool sc_helper_check_arc(sc_memory_context const * ctx, sc_addr beg_el, sc_addr end_el, sc_type arc_type);
 
-/*! Check if current version equal to specified
- * @return If version equal to current memory version, then function returns SC_TRUE;
- * otherwise it returns SC_FALSE
- */
-_SC_EXTERN sc_bool sc_helper_check_version_equal(sc_uint8 major, sc_uint8 minor, sc_uint8 patch);
-
-#endif  // IDENTIFICATION_H
+#endif  // _sc_helper_h_