+++ sc-tools/sc-server/src/sc-server-impl/sc_server_action.hpp
@@ -1,0 +1,24 @@
+#pragma once
+
+#include "sc_server_defines.hpp"
+
+class ScServerAction
+{
+public:
+  explicit ScServerAction(ScServerConnectionHandle hdl)
+    : m_hdl(std::move(hdl))
+  {
+  }
+
+  virtual void Emit() = 0;
+
+  ScServerConnectionHandle GetHandle()
+  {
+    return m_hdl;
+  }
+
+  virtual ~ScServerAction() = default;
+
+protected:
+  ScServerConnectionHandle m_hdl;
+};