+++ tools/builder/tests/units/test_defines.hpp
@@ -1,0 +1,1 @@
+#define BUILDER_REPO_PATH "/home/nikita/ostis-apps/develop/cim-models-storage-module/ostis-web-platform/sc-machine/bin/sc-builder-test-repo"