+++ docs/changelog.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - United config file
 
 ### Changed
+- Unify sc-builder and sc-server config and params usage
+- Change sc-machine config file structure
 - Fix gwf2scs-translator. Add opportunity to parse russian identifiers
 - Simplify main CMakeLists.txt
 - Add tools macros for main CMakeLists.txt